requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytz==2022.2.1
sqlparse==0.4.4
tzdata==2022.2
